pyproject.toml
[project]
name = "piper-dev"
version = "0.1.0"
description = "A Project to Deploy Agliex piper Arm with Orbbec335 Camera"
readme = "README.md"
authors = [
    { name = "Tong Liu", email = "szliutong@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "python-can",
    "numpy<2.0",
    "termcolor>=3.1.0",
    "pybind11==2.11.0",
    "pybind11-global==2.11.0",
    "opencv-python",
    "wheel",
    "av",
    "pygame",
    "pynput",
    "piper-sdk @ git+https://github.com/agilexrobotics/piper_sdk.git@master",
    "pyorbbecsdk @ file:///home/szliutong/Project/piper_dev/whl/pyorbbecsdk-2.0.13-cp310-cp310-linux_x86_64.whl",
    "pillow>=11.3.0",
    "tfds-nightly>=4.9.9.dev202510120044",
    "envlogger>=1.2",
    "tensorflow>=2.18.1",
]

[project.scripts]
piper-dev = "piper_dev:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
